requirements.txt
aiogram==2.7
aiogram-calendar==0.1.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
fake-useragent==0.1.11
frozenlist==1.2.0
gino==1.0.1
idna==3.3
loguru==0.5.3
lxml==4.6.4
multidict==5.2.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
